% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt_generate_exercise_and_solution.r
\name{generate_exercise_and_solution}
\alias{generate_exercise_and_solution}
\title{Generate exercise file and solution file}
\usage{
generate_exercise_and_solution(file_name, compile = TRUE)
}
\arguments{
\item{file_name}{Character string naming a .Rmd file}

\item{compile}{Should generated documents be compiled?}
}
\description{
Generate exercise file and solution file
}
