% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gensolex_main.r
\name{gensolex}
\alias{gensolex}
\title{Generate exercise file and solution file}
\usage{
gensolex(file_name, compile = TRUE)
}
\arguments{
\item{file_name}{Character string naming a .Rmd file}

\item{compile}{Should generated documents be compiled?}
}
\description{
Generate exercise file and solution file
}
\examples{

tf <- system.file("example", "iris.rmd", package = "gensolex")

}
